[Consensus]
- Add channels on blockchain struct for new wire messages
- Change votes from int to uint64
- Move duplicate vote checks before message processing in block reduction and binary agreement
- Dont return blockhash from processMsg functions anymore, as this is already in the message
- Add SignatureSet field to context
- Add preliminary signature set reduction code
- NOTE: signature set reduction code has no tests and is still in progress

+++ pkg/core/consensus/sortition.go
@@ -31,7 +31,7 @@ func sortition(ctx *Context, role *role) error {
 	return nil
 }
 
-func verifySortition(ctx *Context, score, pk []byte, role *role, stake uint64) (int, error) {
+func verifySortition(ctx *Context, score, pk []byte, role *role, stake uint64) (uint64, error) {
 	valid, err := verifyScore(ctx, score, pk, role)
 	if err != nil {
 		return 0, err
@@ -66,7 +66,7 @@ func verifyScore(ctx *Context, score, pk []byte, role *role) (bool, error) {
 	return true, nil
 }
 
-func calcVotes(threshold, stake, totalStake uint64, score []byte) (int, error) {
+func calcVotes(threshold, stake, totalStake uint64, score []byte) (uint64, error) {
 	if threshold > totalStake {
 		return 0, errors.New("threshold size should not exceed maximum stake weight")
 	}
@@ -92,7 +92,7 @@ func calcVotes(threshold, stake, totalStake uint64, score []byte) (int, error) {
 	}
 
 	pos := float64(0.0)
-	votes := 0
+	votes := uint64(0)
 	for {
 		p1 := dist.Prob(pos)
 		p2 := dist.Prob(pos + 0.01)
